+++ docs/ReleaseNotes.md
@@ -1,6 +1,7 @@
 // Before any commit, use the command line: npx tsc --noEmit
 
-# 0.1.9.0 Camera Layer Implementation
+# 0.1.9.1 Refactor and UI changes
 
-- Now sync files are created during project folder creation.
-- Added a Camera Layer, Camera Service, and basic access to camera.
+- Loading component from <Layout /> changed back to scale 1
+- Buttons from photo preview component abstracted
+- Loading feedback from photo preview component abstracted